added stub for cannon and reorganized default baseline compiler
