+++ src/baseline.rs
@@ -1,6 +1,7 @@
 pub use crate::baseline::codegen::{generate, generate_fct};
 
 mod asm;
+pub mod cannon;
 pub mod codegen;
 pub mod dora_compile;
 pub mod dora_entry;
